requirements.txt
discord.py==1.5.1
asyncio==3.4.3
praw==7.1.0
python-dotenv==0.15.0
pathlib==1.0.0